move mustache compatibility tests to its own file
